Refactor logging initialization and add compatibility checks for plugins
