+++ Tunny/Component/LoadingInstruction/InitializeTLog_Tunny.cs
@@ -1,0 +1,24 @@
+using System;
+
+using Grasshopper.Kernel;
+
+using Tunny.Core.Util;
+
+namespace Tunny.Component.LoadingInstruction
+{
+    public class InitializeTLog_Tunny : GH_AssemblyPriority
+    {
+        public override GH_LoadingInstruction PriorityLoad()
+        {
+            try
+            {
+                TLog.InitializeLogger();
+            }
+            catch (Exception)
+            {
+            }
+
+            return GH_LoadingInstruction.Proceed;
+        }
+    }
+}
